requirements.txt
Django==3.1.1
django_rest_knox==4.1.0
django_rest_passwordreset==1.1.0
djangorestframework==3.12.1
gunicorn==20.0.4
whitenoise==5.2.0
django_heroku==0.3.1
